+++ coreexample/src/main/kotlin/me/proton/android/core/coreexample/db/AppDatabaseMigrations.kt
@@ -86,4 +86,11 @@ object AppDatabaseMigrations {
             ContactDatabase.MIGRATION_0.migrate(database)
         }
     }
+
+    val MIGRATION_9_10 = object : Migration(9, 10) {
+        override fun migrate(database: SupportSQLiteDatabase) {
+            AddressDatabase.MIGRATION_2.migrate(database)
+            PublicAddressDatabase.MIGRATION_1.migrate(database)
+        }
+    }
 }