+++ key/data/src/main/kotlin/me/proton/core/key/data/api/response/SignedKeyListResponse.kt
@@ -16,20 +16,15 @@
  * along with ProtonCore.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import studio.forface.easygradle.dsl.*
+package me.proton.core.key.data.api.response
 
-plugins {
-    id("com.android.library")
-    kotlin("android")
-}
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
 
-libVersion = Version(1, 15, 6)
-
-android()
-
-dependencies {
-    api(
-        project(Module.keyDomain),
-        project(Module.keyData)
-    )
-}
+@Serializable
+data class SignedKeyListResponse(
+    @SerialName("Data")
+    val data: String,
+    @SerialName("Signature")
+    val signature: String
+)